chore(zero-cache): structure catchup backlog benchmark
