+++ packages/zero-cache/bench/catchup-backlog/index.ts
@@ -1,0 +1,6 @@
+/* oxlint-disable no-console */
+import {runHandoffBenchmark} from './handoff.ts';
+import {formatReport} from './report.ts';
+
+const results = await runHandoffBenchmark();
+console.log(formatReport(results));
