+++ packages/zero-cache/package.json
@@ -24,7 +24,7 @@
     "bench": "vitest run --config vitest.config.bench.ts",
     "bench:bmf": "BENCH_OUTPUT_FORMAT=json pnpm run bench 2>&1 | node ../shared/src/tool/mitata-json-to-bmf.ts",
     "bench:bmf:silent": "pnpm run bench:bmf --silent",
-    "perf:catchup-backlog": "node ./bench/catchup-backlog.ts",
+    "perf:catchup-backlog": "node ./bench/catchup-backlog/index.ts",
     "local": "node tool/local.ts",
     "start": "node ./src/server/runner/main.ts",
     "benchdb": "node ./bench/bench.ts",
